make playground for all schemas
previously the testing page was just for one schema since
the application itself was for one schema but since multiple
schemas support is added then this page can be used by everyone
